typo fix

+++ src/lib/math.js
@@ -236,8 +236,8 @@ const $builtinmodule = function (name) {
     methods.isclose = function isclose(args, kwargs) {
         Sk.abstr.checkArgsLen("isclose", args, 2, 2);
         rel_abs_vals = Sk.abstr.copyKeywordsToNamedArgs("isclose", ["rel_tol", "abs_tol"], [], kwargs, [
-            new Sk.builtin.float(1e-9),
-            new Sk.builtin.float(0.0),
+            new Sk.builtin.float_(1e-9),
+            new Sk.builtin.float_(0.0),
         ]);
 
         const a = args[0];
